# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PeeZ Neighborhood Membership API
  version: 1.0.0
  description: |
    REST API for PeeZ, a neighborhood membership platform for Oran, Algeria.
    
    ## Business Rule
    Exactly **ONE partner shop** per (neighborhood, category) combination.
    
    ## Authentication
    - Bearer token via Laravel Sanctum
    - Vendor/Admin endpoints require authentication
    - Public endpoints (neighborhoods, categories, shops) accessible without auth
    
    ## Data Formats
    - JSON only
    - Timestamps: ISO-8601 UTC
    - Request/Response: camelCase
    - Database: snake_case
    
    ## Rate Limiting
    60 requests per minute per token
    
  contact:
    name: PeeZ API Support
    email: api@peez.dz
  license:
    name: Proprietary
servers:
  - url: https://api.peez.dz/api/v1
    description: Production
  - url: http://localhost/api/v1
    description: Development
...
security:
  - bearerAuth: []
paths:
...
  /shops/{id}:
    get:
      summary: Get shop details
      tags: [Public]
      security: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Shop details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Shop'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== USER ENDPOINTS ====================
...
  /admin/shops/{id}:
    get:
      summary: Get shop (admin)
      tags: [Admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Shop details
    
    put:
      summary: Update shop
      tags: [Admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                discountPercent:
                  type: number
                  minimum: 5.0
                  maximum: 8.0
                phone:
                  type: string
                isActive:
                  type: boolean
      responses:
        '200':
          description: Shop updated
    
    delete:
      summary: Delete shop
      tags: [Admin]
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Shop deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Neighborhood:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        city:
          type: string
          default: Oran
    Category:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
    ShopSummary:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        discountPercent:
          type: number
          format: float
        avgRating:
          type: number
          format: float
          nullable: true
        ratingsCount:
          type: integer
        lat:
          type: number
          format: double
        lng:
          type: number
          format: double
        neighborhood:
          $ref: '#/components/schemas/Neighborhood'
        category:
          $ref: '#/components/schemas/Category'
    Shop:
      allOf:
        - $ref: '#/components/schemas/ShopSummary'
        - type: object
          properties:
            phone:
              type: string
            isActive:
              type: boolean
...
    ProblemDetails:
      type: object
      description: RFC7807 Problem Details
      properties:
        type:
          type: string
          format: uri
          description: URI reference to problem type
        title:
          type: string
          description: Short human-readable summary
        status:
          type: integer
          description: HTTP status code
        detail:
          type: string
          description: Human-readable explanation
        code:
          type: string
          description: Application-specific error code
        key:
          type: string
          description: Field or parameter that caused the error
      required:
        - type
        - title
        - status
  responses:
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
          example:
            type: "https://api.peez.dz/problems/not-found"
            title: "Resource Not Found"
            status: 404
            detail: "The requested resource does not exist"
            code: "NOT_FOUND"
...